Move AssertNoQtLogWarnings into separate header
This helper to bridge the gap between googletest and
QAbstractItemModelTester shall be used in another item model as well.
